refactor(protocol): derive root types from schema
The maintainer approved broadening the additive schema-backed type surface without a protocol version bump.
