+++ packages/gateway-protocol/src/index.ts
@@ -34,4 +34,19 @@ export * from "./migration-api.js";
 export type * from "./public-session-catalog.js";
 export * from "./validator-registry.js";
 export * from "./schema-export-registry.js";
-export type * from "./type-export-registry.js";
+export type * from "./schema.js";
+
+// Local structural result keeps this package independent of core session types.
+export type SessionsPatchResult = {
+  ok: true;
+  path: string;
+  key: string;
+  entry: Record<string, unknown>;
+  resolved?: {
+    modelProvider?: string;
+    model?: string;
+    agentRuntime?: import("./schema.js").GatewayAgentRuntime;
+    thinkingLevel?: string;
+    thinkingLevels?: Array<{ id: string; label: string }>;
+  };
+};
